requirements.txt
# ============================================
# OpenClaw Sentiment Analysis - Python Dependencies
# ============================================

# Core Framework
pyspark==3.5.0
kafka-python==2.0.2

# Data Processing
pandas==2.0.3
numpy==1.24.3

# Machine Learning
scikit-learn==1.3.0

# Database
pyodbc==5.0.1
pymssql==2.2.11

# API Clients
praw==7.7.1
python-dotenv==1.0.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1

# Utilities
requests==2.31.0